requirements.txt
uvicorn~=0.34.0
fastapi~=0.115.11
pydantic~=2.10.6
passlib~=1.7.4
pdfkit~=1.0.0
python-jose~=3.4.0
Markdown~=3.7
PyYAML~=5.1
frontmatter~=3.0.8
jwt~=1.3.1
bcrypt~=4.3.0